requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
web3==6.16.0